Adding growth tests for RT model.
Added gene ess. tests (for M9 minimal media).
There is one minor issue, and a lot of todos. First, genes in this model are stored as str, but parsed at int (10001), where it should be G_1000. So when loading in the ess. data, it fails. No way to fix without hard coding memote or changing model gene names (probably change gene names).